move skimage-specific code out of _vendored submodule and into skimage/filters/rank
